requirements.txt
fastai==2.7.18
fastapi==0.115.2
Pillow==11.0.0
pydantic==2.9.2
uvicorn==0.32.0
